Address code review: payload over static values, DB-owned timestamps
Review findings on top of the API-semantics fixes. Grouped by what they
actually are rather than by file.

Correctness

- Customer.active was never set and the entity field is a primitive, so
  every customer created through the API was written with active = 0 and
  the column's DEFAULT TRUE never applied. CustomerDTO.active is now a
  boxed Boolean: absent means the default, an explicit value wins. This is
  the same "stop substituting static values for the payload" fix as the
  store/address one, which had left the flag behind.
- Film.filmId was widened to Integer but the four other mappings of the
  same SMALLINT UNSIGNED column -- FilmCategoryId, FilmText, FilmListId,
  NicerButSlowerFilmListId -- were still short/Short, so the >32767
  overflow survived on those paths. All Integer now.
- PATCH /actors/{id} silently discarded an actorId in the body that
  disagreed with the URL, relying on the next line putting it back, while
  POST rejects a client-supplied id outright. It now answers 400.

Timestamps, one rule instead of two

Every last_update column in Sakila is TIMESTAMP NOT NULL DEFAULT
CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP, so all fifteen now map as
insertable = false, updatable = false and no business code sets them.
Previously Actor and Film used insertable = false while FilmActor and
Customer wrote the value by hand, which meant the rule CLAUDE.md stated
was already false in-repo and the next entity with a create() would have
repeated the MySQL 8 error 1048. updatable = false matters too: left
updatable, Hibernate writes back the stale loaded value and suppresses
ON UPDATE. Verified live -- last_update advances across a PATCH with no
Java writing it. customer.create_date is the one column with no default,
so save() still sets it.

This also removes the mixed java.util.Date / LocalDateTime writes.

Test environment

The container forced serverTimezone=UTC while the app now uses
America/Lima. Connector/J converts TIMESTAMP through the connection zone,
so tests and dev disagreed by the offset and a timestamp bug could only
show up in one of them. Both are Lima now, and TZ is set on the container
to match the documented docker line.

Tests

- replaceCustomer permanently rewrote fixture customer 1 to
  INTEGRATION/REPLACED while readCustomer asserts it is still MARY; only
  JUnit's method ordering kept them apart. It creates and deletes its own
  row now.
- createCustomer and createCustomerHonoursPayloadStoreAndAddress issued
  the identical request and differed only in assertions. Merged, plus the
  active flag, plus a case for an explicit active = false.

API consistency

DELETE answered 204 for customers and 200 for actors. All 204 now -- the
same class of defect as the PUT 201 these branches came to fix.

Docs

CLAUDE.md reframed around what is actually a requirement: MySQL 8 is the
hard constraint and the reason is spelled out; port, user and password are
configuration and say so. Adds the caching_sha2_password/useSSL trap for
anyone who sets a real password, the TZ-vs-serverTimezone note, and the
two conventions above. Dropped the stale note in actors.http that still
described the PATCH bug as current behaviour directly above the request
asserting it is fixed.

Not changed: the review also reported http/README.md understating the test
count as 33. It is 33 -- that count included @TestPropertySource as a test.

+++ src/main/java/org/sanmarcux/samples/sakila/dao/model/FilmCategoryId.java
@@ -12,23 +12,23 @@ import jakarta.persistence.Embeddable;
 public class FilmCategoryId implements java.io.Serializable {
 
     private static final long serialVersionUID = 1L;
-    private short filmId;
+    private Integer filmId;
     private byte categoryId;
 
     public FilmCategoryId() {
     }
 
-    public FilmCategoryId(short filmId, byte categoryId) {
+    public FilmCategoryId(Integer filmId, byte categoryId) {
         this.filmId = filmId;
         this.categoryId = categoryId;
     }
 
     @Column(name = "film_id", nullable = false)
-    public short getFilmId() {
+    public Integer getFilmId() {
         return this.filmId;
     }
 
-    public void setFilmId(short filmId) {
+    public void setFilmId(Integer filmId) {
         this.filmId = filmId;
     }
